+++ src/K4os.Compression.LZ4.Streams/ILZ4FrameInfo.cs
@@ -8,7 +8,6 @@
 		bool Chaining { get; }
 		bool BlockChecksum { get; }
 
-		uint? Dictionary { get; }
 		int BlockSize { get; }
 	}
 }