+++ src/K4os.Compression.LZ4.Streams/LZ4FrameInfo.cs
@@ -6,19 +6,17 @@
 		public bool ContentChecksum { get; }
 		public bool Chaining { get; }
 		public bool BlockChecksum { get; }
-		public uint? Dictionary { get; }
 		public int BlockSize { get; }
 
 		public LZ4FrameInfo(
 			long? contentLength, bool contentChecksum,
 			bool chaining, bool blockChecksum,
-			uint? dictionary, int blockSize)
+			int blockSize)
 		{
 			ContentLength = contentLength;
 			ContentChecksum = contentChecksum;
 			Chaining = chaining;
 			BlockChecksum = blockChecksum;
-			Dictionary = dictionary;
 			BlockSize = blockSize;
 		}
 	}