+++ CHANGES.md
@@ -1,5 +1,7 @@
-## 1.0.1-beta (2018/10/03)
+## 1.0.2 (2018/10/03)
 * updated package information
+* added Position to Decode stream
+* added Length to Decode stream, if known
 
 ## 1.0.0-beta (2018/09/09)
 * based on lz4 1.8.1